Fix sample projects integration tests stability (#5489)
* dummy change

* enhance logging in ComponentService

Example:
java.lang.IllegalArgumentException: Invalid result of the Component file [PaymentComponent.js] in project [sample-component-decorators]. The reference of the component class must be the last expression in the file. Returned object is: [JavaScript source code for project name 'sample-component-decorators' and file name 'PaymentComponent.js' could not be found, consider publishing it.] of type [class java.lang.String] but expected instance of class org.graalvm.polyglot.Value
	at org.eclipse.dirigible.components.engine.di.service.ComponentService.executeJavaScript(ComponentService.java:65) ~[classes/:na]

* componnets - allow processing of failed artifacts

sometimes the first processing fails because the transpiled js file is missing at the moment of synchronization

* enhance logging

* fix deplete logic

The previous logic was trying to process the left artifacts over and over again until the retry count is reached.